removed interface, just use class
